fix-ci: extract comparison dispatch into utilities to cut duplication
The quality gate failed on 6.3% duplicated new code (limit 3%) because
S9148 and S9354 still shared the same visitNode shape after dropping the
abstract check class. ComparisonMethodUtils now owns that dispatch.
