+++ java-checks/src/main/java/org/sonar/java/checks/FloatingPointComparisonCheck.java
@@ -16,18 +16,13 @@
  */
 package org.sonar.java.checks;
 
-import java.util.Arrays;
 import java.util.List;
 import org.sonar.check.Rule;
 import org.sonar.java.checks.helpers.ComparisonMethodUtils;
 import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
 import org.sonar.plugins.java.api.semantic.Type;
-import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
 import org.sonar.plugins.java.api.tree.BinaryExpressionTree;
-import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
-import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
-import org.sonar.plugins.java.api.tree.MethodTree;
 import org.sonar.plugins.java.api.tree.Tree;
 
 @Rule(key = "S9148")
@@ -37,25 +32,14 @@ public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
 
   @Override
   public List<Tree.Kind> nodesToVisit() {
-    return Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION);
+    return ComparisonMethodUtils.nodesToVisit();
   }
 
   @Override
   public void visitNode(Tree tree) {
-    if (context.getSemanticModel() == null) {
-      return;
-    }
-    if (tree.is(Tree.Kind.METHOD)) {
-      MethodTree methodTree = (MethodTree) tree;
-      if (ComparisonMethodUtils.isCompareMethod(methodTree)) {
-        methodTree.block().accept(new FloatingPointComparisonVisitor());
-      }
-    } else {
-      LambdaExpressionTree lambda = (LambdaExpressionTree) tree;
-      if (ComparisonMethodUtils.isComparatorLambda(lambda)) {
-        lambda.body().accept(new FloatingPointComparisonVisitor());
-      }
-    }
+    ComparisonMethodUtils.visitComparisonNode(context, tree,
+      methodTree -> methodTree.block().accept(new FloatingPointComparisonVisitor()),
+      lambda -> lambda.body().accept(new FloatingPointComparisonVisitor()));
   }
 
   private static boolean hasFloatingType(ExpressionTree tree) {
@@ -63,7 +47,7 @@ public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
       || tree.symbolType().isPrimitive(Type.Primitives.DOUBLE);
   }
 
-  private class FloatingPointComparisonVisitor extends BaseTreeVisitor {
+  private class FloatingPointComparisonVisitor extends ComparisonMethodUtils.SkipNestedTypesVisitor {
 
     @Override
     public void visitBinaryExpression(BinaryExpressionTree tree) {
@@ -77,16 +61,6 @@ public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
 
     private boolean hasFloatingOperand(BinaryExpressionTree tree) {
       return hasFloatingType(tree.leftOperand()) || hasFloatingType(tree.rightOperand());
-    }
-
-    @Override
-    public void visitClass(ClassTree tree) {
-      // Do not visit inner classes
-    }
-
-    @Override
-    public void visitLambdaExpression(LambdaExpressionTree tree) {
-      // Do not visit nested lambdas
     }
   }
 }
