+++ java-checks/src/main/java/org/sonar/java/checks/helpers/ComparisonMethodUtils.java
@@ -16,9 +16,16 @@
  */
 package org.sonar.java.checks.helpers;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Consumer;
+import org.sonar.plugins.java.api.JavaFileScannerContext;
 import org.sonar.plugins.java.api.semantic.MethodMatchers;
+import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
+import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
+import org.sonar.plugins.java.api.tree.Tree;
 
 public final class ComparisonMethodUtils {
 
@@ -37,11 +44,45 @@ public final class ComparisonMethodUtils {
   private ComparisonMethodUtils() {
   }
 
+  public static List<Tree.Kind> nodesToVisit() {
+    return Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION);
+  }
+
+  public static void visitComparisonNode(JavaFileScannerContext context, Tree tree,
+    Consumer<MethodTree> onCompareMethod, Consumer<LambdaExpressionTree> onComparatorLambda) {
+    if (context.getSemanticModel() == null) {
+      return;
+    }
+    if (tree.is(Tree.Kind.METHOD)) {
+      MethodTree methodTree = (MethodTree) tree;
+      if (isCompareMethod(methodTree)) {
+        onCompareMethod.accept(methodTree);
+      }
+    } else {
+      LambdaExpressionTree lambda = (LambdaExpressionTree) tree;
+      if (isComparatorLambda(lambda)) {
+        onComparatorLambda.accept(lambda);
+      }
+    }
+  }
+
   public static boolean isCompareMethod(MethodTree methodTree) {
     return methodTree.block() != null && COMPARE_METHODS.matches(methodTree);
   }
 
   public static boolean isComparatorLambda(LambdaExpressionTree lambda) {
     return lambda.symbolType().isSubtypeOf("java.util.Comparator");
+  }
+
+  public static class SkipNestedTypesVisitor extends BaseTreeVisitor {
+    @Override
+    public void visitClass(ClassTree tree) {
+      // Do not visit inner classes
+    }
+
+    @Override
+    public void visitLambdaExpression(LambdaExpressionTree tree) {
+      // Do not visit nested lambdas
+    }
   }
 }
